[AnLDD][Change posititon function from role Storekeeper Update Stock/Lot  to role Admin ]
