+++ BL3w_GroupProject_WareHourse/BL3w_GroupProject/Pages/Admin/ManageStockOut/StockOutEdit.cshtml
@@ -1,17 +1,17 @@
 ﻿@page
-@model BL3w_GroupProject.Pages.StoreKeeper.StockOutEditModel
+@model BL3w_GroupProject.Pages.Admin.ManageStockOut.StockOutEditModel
 <head>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
 </head>
 @{
     ViewData["Title"] = "StockOutEdit";
-    Layout = "_StoreKeeper";
+    Layout = "/Pages/Shared/_LayoutAdmin.cshtml";
     int x = 0;
 }
 
 <h1>Stock Out Edit</h1>
 
-<hr/>
+<hr />
 <div class="row">
     <div class="col-md-4">
         <form method="post">
@@ -29,7 +29,7 @@
             </div>
             <div class="form-group">
                 <label asp-for="StockOut.Note" class="control-label"></label>
-                <textarea asp-for="StockOut.Note" class="form-control" ></textarea>
+                <textarea asp-for="StockOut.Note" class="form-control"></textarea>
                 <span asp-validation-for="StockOut.Note" class="text-danger"></span>
             </div>
             <hr />
@@ -38,7 +38,8 @@
                     <tr>
                         <th>No.</th>
                         <th width="250px">Product Name</th>
-                        <th>Quantity<th>
+                        <th>Quantity
+                        <th>
                     </tr>
                 </thead>
                 <tbody>
@@ -65,7 +66,7 @@
 
             <br />
             <div class="form-group">
-                <a asp-page="StockOutList" class="btn text-white" style="background-color: #F37021"><i class="fa-solid fa-right-from-bracket"></i></a>
+                <a asp-page="ListStockOut" class="btn text-white" style="background-color: #F37021"><i class="fa-solid fa-right-from-bracket"></i></a>
                 <input type="submit" value="Save" class="btn btn-primary" />
             </div>
         </form>
@@ -75,5 +76,7 @@
 </div>
 
 @section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
 }
