+++ BL3w_GroupProject_WareHourse/BL3w_GroupProject/Pages/StoreKeeper/LotList.cshtml
@@ -101,9 +101,6 @@
                 <td>
                     <a asp-page="./LotDetail" asp-route-id="@item.LotId"><i class="fa-solid fa-circle-info"></i></a>
                 </td>
-                <td>
-                    <a asp-page="./LotEdit" asp-route-id="@item.LotId"><i class="fa-solid fa-pen"></i></a>
-                </td>
             </tr>
         }
     </tbody>
